+++ api/models/like.go
@@ -8,9 +8,9 @@ import (
 
 // Like A user like a post
 type Like struct {
-	ID        uint      `gorm:"primary_key;auto_increment" json:"id"`
-	PostID    uint      `gorm:"not null;index:idx_post_id" json:"post_id"`
-	UserID    uint      `gorm:"not null;index:idx_user_id" json:"user_id"`
+	ID        uint64    `gorm:"primary_key;auto_increment" json:"id"`
+	PostID    uint64    `gorm:"not null;index:idx_post_id" json:"post_id"`
+	UserID    uint64    `gorm:"not null;index:idx_user_id" json:"user_id"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 	Deleted   gorm.DeletedAt
