+++ api/models/post.go
@@ -8,14 +8,14 @@ import (
 
 // Post post struct
 type Post struct {
-	ID         uint       `gorm:"primary_key;auto_increment" json:"id"`
+	ID         uint64     `gorm:"primary_key;auto_increment" json:"id"`
 	Title      string     `gorm:"size:256;not null;index:idx_title" json:"title"`
 	Summary    string     `gorm:"not null" json:"summary"`
 	Content    string     `gorm:"not null" json:"content"`
 	Categories []Category `gorm:"not null;many2many:post_categories" json:"categories"`
 	Comments   []Comment  `json:"comments"`
 	Likes      []Like     `json:"likes"`
-	UserID     uint       `gorm:"not null;index:idx_user_id" json:"user_id"`
+	UserID     uint64     `gorm:"not null;index:idx_user_id" json:"user_id"`
 	CreatedAt  time.Time  `json:"created_at"`
 	UpdatedAt  time.Time  `json:"updated_at"`
 	Deleted    gorm.DeletedAt
