+++ api/models/user.go
@@ -8,7 +8,7 @@ import (
 
 // User represents a user.
 type User struct {
-	ID        uint      `gorm:"primary_key;auto_increment" json:"id"`
+	ID        uint64    `gorm:"primary_key;auto_increment" json:"id"`
 	Nickname  string    `gorm:"size:256;not null;uniqueIndex:idx_nickname;" json:"nickname"`
 	Type      uint8     `gorm:"not null;index:idx_type" json:"type"`
 	Mobile    string    `gorm:"size:16;not null;uniqueIndex:idx_mobile;" json:"mobile"`
